social-autopilot/pyproject.toml
[project]
name = "social-autopilot"
version = "0.1.0"
description = "Multi-agent LangGraph pipeline for Social Life on Auto-Pilot"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-ollama>=1.0.1",
]

[project.scripts]
social-autopilot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]
include = ["*.py"]
